Implementation of TagType (#51)
* Implementation of TagType

Includes the defaulting of TagType to Stun to allow for ease of tree replacement
Also includes a script for converting old trees

* Fixes an issue with players reporting times in the future

* Added tools for easing tag changes

* Added more functionality to tag backend

* Update admin.py

Added OZ searching and type listing for ease of use

* Anti-similar-tag

This is an attempt to prevent players from submitting a kill twice retroactively and having it slip past the mod team unnocited. Also prevents players from having two people submit the same kill. Needs to be reviewed.

* Update moderator.py

Can't extended a list with a set of tags, so this works instead

* Update player.py

Updated tagtypes to the new tree. Instituted a way for tag descriptions to appear on hovering over edges

Co-authored-by: UW HvZ <40743370+uwhvz@users.noreply.github.com>